llm/requirements.txt
Django>=5.0
openai>=1.0.0
python-dotenv>=1.0.0
google-genai==1.65.0
